requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
streamlit
requests
sentence-transformers
scikit-learn
pandas
plotly
torch
transformers
